remove the change to float16 according to https://github.com/Fannovel16/ComfyUI-Frame-Interpolation/pull/52#issuecomment-1949715465

+++ vfi_models/rife/__init__.py
@@ -102,6 +102,6 @@ class RIFE_VFI:
         args = [interpolation_model, scale_list, fast_mode, ensemble]
         out = postprocess_frames(
             generic_frame_loop(frames, clear_cache_after_n_frames, multiplier, return_middle_frame, *args, 
-                               interpolation_states=optional_interpolation_states, dtype=torch.float16)
+                               interpolation_states=optional_interpolation_states, dtype=torch.float32)
         )
         return (out,)
